TS Type Checking
Going to work on adding TS support to this library!

First I just have to add type checking to the project itself, which is what this commit is for.

Thankfully it seems like a lot of the project both uses modern syntax, as well as has JSDoc typings! That's already half the battle. The other part is de-mangling the use of those types in contexts where TS can't discern things properly yet, say like recursively-defined types, which are just reduced to `any` unfortunately.

```sh
npx tsc --watch
```