Type Sweep
Been working on these commits for the last few hours or so, just now getting to commit them.

The stages I've learned for doing TS migrations, thus far, is that you should fix the existing types if there are any currently, and keep things in plain JS/CJS, and only do the type adjustments first. It doesn't break things like the library tests, or the build step.

https://jsdoc.app/tags-param
https://jsdoc.app/tags-callback

https://www.youtube.com/watch?v=wtyOhM-DwfA
https://www.youtube.com/watch?v=HROjcHR07Pg