Revert "TS Type Checking", "Type Sweep"
This reverts commits b7c09a81b6a5c32195f31384a8ce94213c9042d5 and c83ed6ea57d19e8db2a9fc6a8edbeb297d706a79.

I came back to the repo today to find that others started to contribute what I was going to do as well, but much better! There were thankfully a lot of dependency upgrades as well, and I definitely want to make sure to have that as well, that is a really big help to moving things forward into a nice and fancy codebase.